feat: 채널이름 동기화 완료
